DiffFusionServer

Documentation for DiffFusionServer.

DiffFusionServer._INFO_STRING

DiffFusionServer._NO_HTTP_ERROR

DiffFusionServer.api_get_aliases

DiffFusionServer.api_get_info

DiffFusionServer.initial_repository

DiffFusionServer._INFO_STRING — Constant

const _INFO_STRING

An info string with API details.

DiffFusionServer._NO_HTTP_ERROR — Constant

const _NO_HTTP_ERROR

Default return status code.

DiffFusionServer.api_get_aliases — Method

Return all aliases from the repository in body of response.

DiffFusionServer.api_get_info — Method

api_get_info(req::HTTP.Request)

Return an info string about the API as response.

DiffFusionServer.initial_repository — Method

initial_repository()

Create an object repository and initialise with some static objects.